+++ heath/src/app/(needLogin)/main/[postId]/DetailPage.styles.ts
