+++ heath/src/app/(needLogin)/main/_components/PostCard/PostCard.types.ts
@@ -1,0 +1,15 @@
+import { StaticImageData } from 'next/image'
+
+export interface PostCardProps {
+  post: {
+    id: number
+    image: StaticImageData
+    title: string
+    content: string
+    created: Date
+    user: {
+      id: string
+      profileImage: StaticImageData
+    }
+  }
+}
